+++ lib/NumberFormatter.php
@@ -12,26 +12,7 @@ use ICanBoogie\CLDR\Numbers\Symbols;
 class NumberFormatter implements Formatter, Localizable
 {
 	/**
-	 * Format a number with the specified pattern.
-	 *
-	 * Note, if the pattern contains '%', the number will be multiplied by 100 first. If the
-	 * pattern contains '‰', the number will be multiplied by 1000.
-	 *
-	 * @param float|int|numeric-string $number
-	 *     The number to format.
-	 * @param string|NumberPattern $pattern
-	 *     The pattern used to format the number.
-	 */
-	public function __invoke(
-		float|int|string $number,
-		NumberPattern|string $pattern,
-		Symbols $symbols = null,
-	): string {
-		return $this->format($number, $pattern, $symbols);
-	}
-
-	/**
-	 * Format a number with the specified pattern.
+	 * Formats a number with the specified pattern.
 	 *
 	 * Note, if the pattern contains '%', the number will be multiplied by 100 first. If the
 	 * pattern contains '‰', the number will be multiplied by 1000.